move styles to commonStyles
